requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
idna==3.4
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.1
protobuf==3.20.3
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyYAML==6.0
QtAwesome==1.2.3
QtPy==2.3.1
requests==2.30.0
urllib3==2.0.2
